requirements_dev.txt
pytest
requests_mock
pytest-socket